+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Nonce attribute is missing in script tag for TYPO3 v13.4.20+ with activated CSP (#34)
+
 ## [2.4.0] - 2025-09-17
 
 Documentation improvements